major changes to predict.models due to peer review
